requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
colorama==0.4.6
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.28.2
setuptools==69.5.1
